24.06.03
- Mesh작업
- Object 객체 구조 수정
   - 최상위 클래스 Object 추상 클래스로 수정
   - 그 하위 클래스로 Actor 클래스 추가

+++ Framework/Utilities/BinaryFile.cpp
@@ -25,7 +25,7 @@ BinaryWrite::BinaryWrite(string FilePath)
 
 #if DEBUG == 1
 	{
-		int index = FilePath.find_last_of('.');
+		int index = static_cast<int>(FilePath.find_last_of('.'));
 		string debugFilePath = FilePath.substr(0, index);
 		debugFilePath += ".txt";
 		this->DebugFile.open(debugFilePath, std::ios_base::out);
